feat: add dtos and crud operations for invitations

+++ backend/src/invitations/invitations.controller.ts
@@ -1,10 +1,11 @@
 import {
   Body,
+  ConflictException,
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
-  Patch,
   Post,
   Query,
 } from '@nestjs/common';
@@ -12,13 +13,20 @@ import { InvitationsService } from './invitations.service';
 import { CreateInvitationDto } from './dto/create-invitation.dto';
 import { UpdateInvitationDto } from './dto/update-invitation.dto';
 import { SearchDto } from '../common/dto/search.dto';
+import { User } from 'src/auth/user.decorator';
+import { User as UserEntity } from 'src/users/entities/user.entity';
 
 @Controller('invitations')
 export class InvitationsController {
   constructor(private readonly invitationsService: InvitationsService) {}
 
   @Post()
-  create(@Body() createInvitationDto: CreateInvitationDto) {
+  create(
+    @Body() createInvitationDto: CreateInvitationDto,
+    @User() user: UserEntity,
+  ) {
+    createInvitationDto.sender = user;
+
     return this.invitationsService.create(createInvitationDto);
   }
 
@@ -32,16 +40,37 @@ export class InvitationsController {
     return this.invitationsService.findOne(id);
   }
 
-  @Patch(':id')
-  update(
+  @Post(':id/respond')
+  async respond(
     @Param('id') id: string,
     @Body() updateInvitationDto: UpdateInvitationDto,
+    @User() user: UserEntity,
   ) {
-    return this.invitationsService.update(id, updateInvitationDto);
+    const invitation = await this.invitationsService.findOne(id, {
+      receiver: true,
+    });
+
+    if (invitation.receiver.id != user.id) {
+      throw new ForbiddenException();
+    }
+
+    if (invitation.expirationDate < new Date()) {
+      throw new ConflictException();
+    }
+
+    await this.invitationsService.respond(id, updateInvitationDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string, @User() user: UserEntity) {
+    const invitation = await this.invitationsService.findOne(id, {
+      sender: true,
+    });
+
+    if (invitation.sender.id != user.id) {
+      throw new ForbiddenException();
+    }
+
     return this.invitationsService.remove(id);
   }
 }
